+++ gzap/CHANGELOG.md
@@ -3,6 +3,7 @@
 - New:
   - New graphics for checks that still exist in the world but have no items in
     them (e.g. because of quickload or `!collect`).
+- Changed:
   - Changes to default yaml settings:
     - `included_levels` now defaults to all maps in the WAD.
     - `starting_levels` contains all episode-start maps, but may include additional
@@ -10,6 +11,8 @@
     - `start_with_keys` is now off by default if the WAD has tuning data.
     - `start_with_keys` is now forced off for WADs using hub logic, as turning it
       on will start you with all the keys for the entire episode.
+    - `included_item_categories` now excludes secrets and items in secrets by
+      default.
 - Fixed:
   - UT did not work properly with non-default `included_item_categories`.
   - UT indicators in-game did not work reliably.